change coverted value of navigate tree

+++ L09-jsonwriter/src/main/java/ru/otus/l9/jsonwriter/MyJson.java
@@ -157,7 +157,7 @@ public class MyJson {
                     setFieldValue( inputeObject, keyName, parser.getString() );
                     break;
                 case VALUE_NUMBER:
-                    setFieldValue( inputeObject, keyName, parser.getLong() );
+                    setFieldValue( inputeObject, keyName, parser.getInt() );
                     break;
             }
         }
